+++ src/framework/piral-core/src/tools/codegen.ts
@@ -44,7 +44,14 @@ interface CodegenOptions {
   appName: string;
   externals: Array<string>;
   publicPath: string;
-  debug: boolean;
+  debug?: {
+    viewState?: boolean;
+    loadPilets?: boolean;
+    hardRefresh?: boolean;
+    viewOrigins?: boolean;
+    extensionCatalogue?: boolean;
+    clearConsole?: boolean;
+  };
   emulator: boolean;
 }
 
@@ -140,8 +147,9 @@ export function createDebugHandler(imports: Array<string>, exports: Array<string
 
   // if we build the debug version of piral (debug and emulator build)
   if (debug) {
-    imports.push(`import { integrateDebugger } from "piral-core/${cat}/tools/debugger"`);
-    exports.push(`export { integrateDebugger }`);
+    const originalCall = `originalDebugger(context, options, { defaultSettings: ${JSON.stringify(debug)}, ...debug })`;
+    imports.push(`import { integrateDebugger as originalDebugger } from "piral-core/${cat}/tools/debugger"`);
+    exports.push(`export function integrateDebugger(context, options, debug) { return ${originalCall}; }`);
   } else {
     exports.push(`export function integrateDebugger() {}`);
   }